+++ .gitignore
@@ -5,16 +5,6 @@ library
 node_modules
 .tmp
 
-### ESOPS GITIGNORE AUTO GENERATED BEGIN ###
-/.node-dev.json
-/.nycrc
-/.prettierrc
-/.vscode/extensions.json
-/.vscode/settings.json
-/source/readme.md
-/tsconfig.json
-### ESOPS GITIGNORE AUTO GENERATED END ###
-
 ### ESOPS AUTO GENERATED BEGIN ###
 /.node-dev.json
 /.nycrc
