+++ source/drivers/index.ts
@@ -1,7 +1,2 @@
 export {default as fs} from './fs/index'
 export {default as console} from './console/index'
-
-export const throwError = message => e => {
-  console.error(e)
-  throw new Error(message)
-}
