--- source/drivers/process/index.ts
@@ -1,1 +1,0 @@
-export default process
