+++ source/utils/async.ts
@@ -1,0 +1,8 @@
+/**
+ * Async
+ */
+const pipe = require('promised-pipe')
+
+export {pipe}
+
+export {default as result} from 'await-result'
