+++ source/utils/sync.ts
@@ -1,8 +1,5 @@
-import {is, tryCatch as RTryCatch} from 'ramda'
+import {is} from 'ramda'
 
-/**
- * Sync
- */
 export const isString = (arg: any): arg is string => is(String)(arg)
 
 export const throwIfTypeIs = type => arg => {
@@ -14,14 +11,7 @@ export const throwError = e => {
   throw e
 }
 
-export {curry, tryCatch, is} from 'ramda'
-
 // [ramda types filter fix](https://github.com/DefinitelyTyped/DefinitelyTyped/issues/25581#issuecomment-442354314)
 export const filter = fn => arr => arr.filter(fn)
 
-/**
- * Async
- */
-const pipe = require('promised-pipe')
-export {pipe}
-export {default as result} from 'await-result'
+export {curry, tryCatch, is} from 'ramda'
